Swapchain recreation #2
Factored out swapchain configuration.
Refactored surface and swapchain builder accordingly.
Added wrapper for device limits.
Added surface format wrapper tested by equality.

+++ src/main/java/org/sarge/jove/platform/vulkan/memory/Allocator.java
@@ -53,7 +53,7 @@ public class Allocator {
 		final var limits = device.limits();
 		this.device = requireNonNull(device);
 		this.types = Arrays.copyOf(types, types.length);
-		this.page = requireOneOrMore((int) limits.get("bufferImageGranularity"));
+		this.page = requireOneOrMore((long) limits.get("bufferImageGranularity"));
 		this.max = constrain(limits.get("maxMemoryAllocationCount"), Integer.MAX_VALUE);
 	}
